readme
The Applet is capable of generating cryptographic keys on the card, and allows external keys to be inserted onto the card. These keys can be used in cryptographic operations, after proper user (or host application) authentication. 

The Applet is capable of handling generic objects. An object is a sequence of bytes whose meaning is determined by the application. The Applet allows a host application to read and/or modify objects’ contents, after proper user (or host application) authentication.